feat: add PG instrumentation and some refactoring
