+++ packages/app/package.json
@@ -89,8 +89,8 @@
     "@opentelemetry/exporter-trace-otlp-http": "0.57.1",
     "@opentelemetry/instrumentation-express": "0.47.0",
     "@opentelemetry/instrumentation-http": "0.57.1",
+    "@opentelemetry/instrumentation-pg": "0.50.0",
     "@opentelemetry/instrumentation-pino": "0.46.0",
-    "@opentelemetry/propagator-b3": "1.30.1",
     "@opentelemetry/propagator-jaeger": "1.30.1",
     "@opentelemetry/sdk-node": "0.57.1",
     "@opentelemetry/sdk-trace-base": "1.30.1",
